refactor: 옵션 더알아보기 팝업 로직 OptionPage 상단으로 옮김
