+++ front/src/Pages/CustomPage/OptionPage/DefaultOption/index.jsx
@@ -9,18 +9,11 @@ import { options } from "../optionData";
 import left from "../../../../assets/icons/chevron-left.svg";
 import right from "../../../../assets/icons/chevron-right.svg";
 import TaggedPage from "../TaggedPage";
-import OverlaidPopup from "../../../../Components/Common/OverlaidPopup";
-import OptionPopup from "../OptionPopup";
-import useOnClickPopUp from "../../../../hooks/useOnClickPopUp";
 
-const DefaultOption = () => {
+const DefaultOption = ({ openPopup }) => {
   const tagsOption = [tags, tagSelectIcons, tagsNotSelectIcons];
   const [selectTag, setSelectTag] = useState("대표");
   const handleSelectTag = (tag) => setSelectTag(tag);
-
-  const optionPopupRef = useRef();
-  const { isPopupOpen, openPopup, closePopup } =
-    useOnClickPopUp(optionPopupRef);
 
   // 탭
   const [selectOption, setSelectOption] = useState([]);
@@ -69,7 +62,7 @@ const DefaultOption = () => {
     const currentData = getDataForPage(newData, currentPage, 8);
     totalPages = Math.ceil(newData.length / 8);
 
-    return <TaggedPage />;
+    return <TaggedPage openPopup={openPopup} />;
     // return currentData.map((data, index) => (
     //   <OptionItem
     //     key={index}
@@ -88,13 +81,13 @@ const DefaultOption = () => {
 
   return (
     <Wrapper>
-      {isPopupOpen && (
+      {/* {isPopupOpen && (
         <OverlaidPopup
           component={
             <OptionPopup popupRef={optionPopupRef} closePopup={closePopup} />
           }
         />
-      )}
+      )} */}
       <OptionTag
         selectTag={selectTag}
         tagsOption={tagsOption}
